建立了 Menu 頁面，以及參考 Scene Builder 的方式將原始碼做了一些調整。並將原本的自創類別 MyButton 及 Hexagon 各建一個 Java 檔
