+++ hex/src/main/java/javahex/hex/Menu.java
@@ -1,0 +1,16 @@
+package javahex.hex;
+
+import javafx.fxml.FXML;
+
+public class Menu {
+    @FXML
+    public void startBtnClick(){
+        App.hexScene.getRoot().requestFocus();
+        App.stage.setScene(App.hexScene);
+    }
+
+    @FXML
+    public void exitBtnClick(){
+        App.stage.close();
+    }
+}
