+++ hex/src/main/resources/javahex/hex/hex.fxml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.*?>
+<?import javafx.scene.layout.*?>
+<?import javafx.scene.text.*?>
+
+<VBox maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" onKeyTyped="#startHex" prefHeight="400.0" prefWidth="400.0" xmlns="http://javafx.com/javafx/16" xmlns:fx="http://javafx.com/fxml/1" fx:controller="javahex.hex.Hex">
+   <children>
+      <Label contentDisplay="CENTER" text="按下空白開始">
+         <font>
+            <Font size="50.0" />
+         </font>
+      </Label>
+   </children>
+</VBox>
